chatbot-redis-backend/MM2/requirements.txt
fastapi
fastapi[standard]
groq
uvicorn
fastapi-cors
redis[hiredis]
redis
python-dotenv
supabase
requests
openai>=1.0.0
mem0ai
uuid-utils
pinecone[grpc]
fastapi-utils
typing_inspect
pytest

cartesia